User lookup via SOAP now/added group lookup
For SP 2010 compatibility, added SOAP actions and removed the old page
scraping code.  Also changed the lookup caching to session only for now
(will likely change this again later on but this made the code much
simpler).

Also added the ability to load the user’s SP group memberships.  This
is useful to change the view based on the user’s memberships.

Also added xmlToJSON utility function to handle the conversion of the
SOAP responses to JSON.

Finally, added a persistent callback that will immediately return the
user + groups info on subsequent calls.